Move colormap util to new file
